Removing specific version info
The index page is not the best place to contain specific version information about packages contained by product. Those versions can change frequently, so it's likely that this documentation will often be out of date. It's better to restrict this information to a more authoritative location.

I haven't decided on the best place to publish that, so removing the (mis-)information for now. If it makes sense later on, will add a reference to the appropriate place, as we have done with Backups, above.