changing definition of "cpu_time" [#158567736]
Originally defined: "Emitted only by cf-mysql-release. Introduced in MariaDB 5.3." 

'cf-mysql-release' is the BOSH release of the MySQL database as a service. But that's not important for this definition. What's important is that this metric (cpu_time) is only emitted when MariaDB is present, and MariaDB only works with MySQL for PCF v1. Thus I changed the description.