Update example default-collation to have underscores
When using the documentation as a reference for what value to set for the default-collation if the value `utf8-general-ci` is copied and pasted into a configuration block it will fail with the following exception:

```
Server error, status code: 502, error code: 10001, message: Service broker error: The 'default-collation' parameter can only be a lowercase alphanumeric string with underscores and no other special characters
```

It might be more usable for the value in the documentation be set to `utf8_general_ci`.  I am not sure if other references in the doc should also be updated since the value that is required should have underscored and not dashes.